Refactor PageRouter.
